+++ pytorch_lightning/trainer/connectors/accelerator_connector.py
@@ -785,10 +785,6 @@ class AcceleratorConnector:
         return self._parallel_devices
 
     @property
-    def num_processes(self) -> int:
-        return self.devices if self.devices is not None else 1
-
-    @property
     def devices(self) -> int:
         if isinstance(self.strategy, SingleDeviceStrategy):
             return 1
@@ -801,13 +797,6 @@ class AcceleratorConnector:
         if isinstance(self.accelerator, TPUAccelerator):
             return self._tpu_cores  # type: ignore
         return 0
-
-    @property
-    def tpu_id(self) -> Optional[int]:
-        if isinstance(self.accelerator, TPUAccelerator):
-            if isinstance(self._tpu_cores, list):
-                return self._tpu_cores[0]
-        return None
 
     @property
     def num_ipus(self) -> int: