+++ tests/accelerators/test_accelerator_connector.py
@@ -551,7 +551,7 @@ def test_validate_accelerator_and_devices():
 
     trainer = Trainer(accelerator="ddp_cpu", devices=2)
     assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert trainer.num_processes == 2
+    assert trainer.num_devices == 2
 
 
 def test_set_devices_if_none_cpu():
@@ -927,7 +927,6 @@ def test_unsupported_ipu_choice(mock_ipu_acc_avail, monkeypatch):
 def test_devices_auto_choice_cpu(is_ipu_available_mock, is_tpu_available_mock, is_gpu_available_mock):
     trainer = Trainer(accelerator="auto", devices="auto")
     assert trainer.num_devices == 1
-    assert trainer.num_processes == 1
 
 
 @mock.patch("torch.cuda.is_available", return_value=True)